api/requirements.txt
Click==7.0
Flask==1.1.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
python-dotenv==0.11.0
Werkzeug==1.0.0
Flask-SQLAlchemy==2.4.4
psycopg2-binary==2.8.5
flask-marshmallow==0.13.0
marshmallow-sqlalchemy==0.23.1
requests==2.24.0